+++ products/admin_urls.py
@@ -1,7 +1,7 @@
 from django.urls import path
-from products.views import (
+from products.admin_views import (
     big_category_form, small_category_form, product_form, 
-    product_edit, product_delete, ingredient_form,
+    product_edit, ingredient_form,
     big_category_delete, small_category_delete, ingredient_delete
 )
 from products.admin_auth import admin_auth_required, admin_logout, admin_login_view
@@ -19,8 +19,7 @@ urlpatterns = [
     path("small-category/", admin_auth_required(small_category_form), name="admin_small_category_form"),
     path("small-category/<int:category_id>/delete/", admin_auth_required(small_category_delete), name="admin_small_category_delete"),
     path("product/", admin_auth_required(product_form), name="admin_product_form"),
-    path("product/<int:product_id>/edit/", admin_auth_required(product_edit), name="admin_product_edit"),
-    path("product/<int:product_id>/delete/", admin_auth_required(product_delete), name="admin_product_delete"),
+    path("product/<int:product_id>/", admin_auth_required(product_edit), name="admin_product_edit"),
     path("ingredient/", admin_auth_required(ingredient_form), name="admin_ingredient_form"),
     path("ingredient/<int:ingredient_id>/delete/", admin_auth_required(ingredient_delete), name="admin_ingredient_delete"),
     path("logout/", admin_logout, name="admin_logout"),
