+++ products/views.py
@@ -10,17 +10,10 @@ from drf_spectacular.utils import extend_schema, OpenApiParameter
 from drf_spectacular.types import OpenApiTypes
 from django.db.models import Sum, Count, Q, Subquery
 from django.db.models.functions import Coalesce
-from products.models import Product, BigCategory, SmallCategory, ProductIngredient, ProductImage, Ingredient, CategoryProduct
+from products.models import Product, BigCategory, ProductIngredient, ProductImage
 from products.serializers import ProductCreateSerializer, ProductReadSerializer, ProductDetailSerializer, ProductRankingSerializer
 from products.serializers import ProductsListSerializer, CategorySerializer, ProductSearchSerializer, MainSerializer
 from products.utils import record_view, upload_images_to_s3
-
-# 템플릿 뷰를 위한 import 추가
-from django.shortcuts import render, redirect
-from django.contrib import messages
-from django.conf import settings
-from django.http import HttpResponseForbidden
-from products.admin_auth import admin_auth_required
 
 # 제품 등록
 class ProductCreateView(generics.CreateAPIView):
@@ -56,142 +49,411 @@ class ProductDetailView(generics.RetrieveAPIView):
         summary="제품 상세 조회",
         tags=["제품"]
     )
-    def get(self, request, *args, **kwargs):
-        return self.retrieve(request, *args, **kwargs)
 
-# 제품 랭킹 (GET /products/ranking/)
+    def get(self, request, *args, **kwargs):
+
+        product = self.get_object()
+
+        record_view(product)
+
+        # serializer에 request 컨텍스트 전달
+
+        serializer = self.get_serializer(product, context={'request': request})
+
+        return Response(serializer.data)
+
+# 랭킹 카테고리 (메인의 인기 카테고리와 동일 순서)
+class ProductRankingCategoryView(generics.ListAPIView):
+    permission_classes = [AllowAny]
+
+    @extend_schema(
+        summary="랭킹 카테고리",
+        tags=["제품"]
+    )
+    def get(self, request, *args, **kwargs):
+        # 소분류별 지난 30일 조회수 합계 순 상위 소분류 목록
+        today = timezone.now().date()
+        start_date = today - timedelta(days=30)
+
+        from products.models import SmallCategory
+
+        small_with_views = (
+            SmallCategory.objects
+            .exclude(category="전체")
+            .annotate(
+                totalViews=Coalesce(
+                    Sum(
+                        "category_products__product__daily_views__views",
+                        filter=Q(category_products__product__daily_views__date__gte=start_date)
+                    ),
+                    0,
+                )
+            )
+            .order_by("-totalViews", "id")
+        )
+
+        categories_payload = [
+            {"smallCategory": sc.category, "bigCategory": sc.bigCategory.category}
+            for sc in small_with_views
+        ]
+
+        return Response({
+            "topSmallCategories": categories_payload
+        })
+
+# 랭킹
 class ProductRankingView(generics.ListAPIView):
     serializer_class = ProductRankingSerializer
     permission_classes = [AllowAny]
 
+    class RankingPagination(PageNumberPagination):
+        page_size = 10                # 고정 10개
+        page_query_param = "page"     # 페이지 번호만 허용
+        page_size_query_param = None   # 클라이언트가 변경 불가
+        max_page_size = 10
+
+    pagination_class = RankingPagination
+
     def get_queryset(self):
+        period = self.request.query_params.get("period", "daily")
+        category = self.request.query_params.get("category")
+
         today = timezone.now().date()
-        start_date = today - timedelta(days=30)
 
-        # 최근 30일간의 일일 조회수 합계를 기준으로 랭킹
-        queryset = Product.objects.annotate(
-            total_views=Coalesce(
-                Sum('daily_views__views', filter=Q(daily_views__date__gte=start_date)),
-                0
-            )
-        ).order_by('-total_views')[:10]
+        if period == "daily":
+            start_date = today - timedelta(days=1)
+        elif period == "monthly":
+            # 월간: 당일 00시 기준 고정(오늘 제외)
+            start_date = today - timedelta(days=30)
+        else:
+            start_date = today - timedelta(days=30)
 
-        return queryset
+        # monthly의 경우 오늘은 제외하여 00시 기준 고정되도록 함
+        if period == "monthly":
+            queryset = Product.objects.filter(daily_views__date__gte=start_date, daily_views__date__lt=today)
+        else:
+            queryset = Product.objects.filter(daily_views__date__gte=start_date)
+
+        if category and category != "전체":
+            queryset = queryset.filter(category_products__category__category=category)
+
+        ranked = queryset.annotate(totalViews=Sum("daily_views__views")).order_by("-totalViews", "id")[:50]
+
+        # 변동 계산 컨텍스트 준비 (daily: 어제, monthly: 이전 30일)
+        if period == "daily":
+            prev_start = today - timedelta(days=1)
+            prev_end = today
+            prev_qs = Product.objects.filter(daily_views__date__gte=prev_start, daily_views__date__lt=prev_end)
+        else:  # monthly (오늘 제외한 직전 30일 비교)
+            prev_start = today - timedelta(days=60)
+            prev_end = today - timedelta(days=30)
+            prev_qs = Product.objects.filter(daily_views__date__gte=prev_start, daily_views__date__lt=prev_end)
+
+        if category and category != "전체":
+            prev_qs = prev_qs.filter(category_products__category__category=category)
+
+        prev_ranked = list(
+            prev_qs.annotate(totalViews=Sum("daily_views__views")).order_by("-totalViews", "id")[:50].values_list("id", flat=True)
+        )
+
+        prev_ranks = {pid: idx + 1 for idx, pid in enumerate(prev_ranked)}
+
+        current_ranks = {product.id: idx + 1 for idx, product in enumerate(ranked)}
+
+        # 인스턴스 변수에 저장하여 list 메서드에서 사용
+        self.ranking_context = {"period": period, "current_ranks": current_ranks, "prev_ranks": prev_ranks}
+
+        return ranked
+
+    def list(self, request, *args, **kwargs):
+        queryset = self.filter_queryset(self.get_queryset())
+
+        page = self.paginate_queryset(queryset)
+        if page is not None:
+            serializer = self.get_serializer(page, many=True, context={
+                'request': request,
+                **(getattr(self, 'ranking_context', {}))
+            })
+            return self.get_paginated_response(serializer.data)
+
+        serializer = self.get_serializer(queryset, many=True, context={
+            'request': request,
+            **(getattr(self, 'ranking_context', {}))
+        })
+        return Response(serializer.data)
 
     @extend_schema(
-        summary="제품 랭킹 조회",
+        summary="랭킹 조회",
+        parameters=[
+            OpenApiParameter(
+                name="period",
+                type=OpenApiTypes.STR,
+                location=OpenApiParameter.QUERY,
+                description="랭킹 기간 (기본: daily)",
+                required=False,
+                enum=["daily", "monthly"],
+            ),
+            OpenApiParameter(
+                name="category",
+                type=OpenApiTypes.STR,
+                location=OpenApiParameter.QUERY,
+                description="카테고리 이름 (예: 전체, 체지방 관리 등). 기본: 전체",
+                required=False,
+            ),
+        ],
         tags=["제품"]
     )
     def get(self, request, *args, **kwargs):
         return self.list(request, *args, **kwargs)
 
-# 제품 목록 (GET /products/list/)
+# 제품 리스트 (필터/정렬/페이징)
 class ProductListView(generics.ListAPIView):
     serializer_class = ProductsListSerializer
     permission_classes = [AllowAny]
-    pagination_class = PageNumberPagination
 
-    def get_queryset(self):
-        queryset = Product.objects.all()
-        
-        # 검색 파라미터
-        search = self.request.query_params.get('search', None)
-        if search:
-            queryset = queryset.filter(
-                Q(name__icontains=search) | Q(company__icontains=search)
-            )
-        
-        return queryset.order_by('-id')
+    class ListPagination(PageNumberPagination):
+        page_size = 10
+        page_query_param = "page"
+        page_size_query_param = None
+        max_page_size = 10
+
+    pagination_class = ListPagination
 
     @extend_schema(
-        summary="제품 목록 조회",
+        summary="제품 리스트 조회",
+        parameters=[
+            OpenApiParameter(
+                name="bigCategory",
+                type=OpenApiTypes.STR,
+                location=OpenApiParameter.QUERY,
+                description="대분류 카테고리 이름(예: 다이어트 약, 다이어트 보조제, 다이어트 식품 등). 미제공시 모든 제품 조회",
+                required=False,
+            ),
+            OpenApiParameter(
+                name="smallCategory",
+                type=OpenApiTypes.STR,
+                location=OpenApiParameter.QUERY,
+                description="소분류 카테고리 이름(예: 전체, 체지방 관리 등). 기본: 전체",
+                required=False,
+            ),
+            OpenApiParameter(
+                name="sort",
+                type=OpenApiTypes.STR,
+                location=OpenApiParameter.QUERY,
+                description="정렬: monthly_rank | price_asc | price_desc | review_desc (기본: monthly_rank)",
+                required=False,
+                enum=["monthly_rank", "price_asc", "price_desc", "review_desc"],
+            ),
+        ],
         tags=["제품"]
     )
     def get(self, request, *args, **kwargs):
         return self.list(request, *args, **kwargs)
 
-# 카테고리별 제품 조회 (GET /products/category/)
+    def get_queryset(self):
+        bigCategory = self.request.query_params.get("bigCategory")
+        smallCategory = self.request.query_params.get("smallCategory")
+        sort = self.request.query_params.get("sort", "monthly_rank")
+
+        today = timezone.now().date()
+        start_date = today - timedelta(days=30)
+
+        qs = Product.objects.all()
+        
+        # bigCategory가 제공된 경우에만 필터링
+        if bigCategory:
+            qs = qs.filter(category_products__category__bigCategory__category=bigCategory).distinct()
+
+        if smallCategory and smallCategory != "전체":
+            qs = qs.filter(category_products__category__category=smallCategory).distinct()
+
+        if sort == "price_asc":
+            return qs.order_by("price", "id")
+
+        if sort == "price_desc":
+            return qs.order_by("-price", "id")
+
+        if sort == "review_desc":
+            return (
+                qs.annotate(review_count=Count("reviews", distinct=True))
+                .order_by("-review_count", "id")
+            )
+
+        # default: monthly_rank (최근 30일 조회수 합계 기준)
+        # 조회수 없는 상품도 포함(0으로 취급)하여 정렬
+        # 월간 랭킹 기준: 오늘 제외하여 00시 기준 고정
+        return (
+            qs.annotate(
+                totalViews=Coalesce(
+                    Sum(
+                        "daily_views__views",
+                        filter=Q(daily_views__date__gte=start_date) & Q(daily_views__date__lt=today),
+                    ),
+                    0,
+                )
+            )
+            .order_by("-totalViews", "id")
+        )
+
+# 제품 카테고리 조회
 class ProductCategoryView(generics.ListAPIView):
     serializer_class = CategorySerializer
     permission_classes = [AllowAny]
 
-    def get_queryset(self):
-        return BigCategory.objects.prefetch_related('smallCategories__category_products__product').all()
-
     @extend_schema(
-        summary="카테고리별 제품 조회",
+        summary="카테고리 조회",
         tags=["제품"]
     )
     def get(self, request, *args, **kwargs):
         return self.list(request, *args, **kwargs)
 
-# 제품 검색 (GET /products/search/)
+    def get_queryset(self):
+        return BigCategory.objects.prefetch_related("smallCategories").order_by("id")
+
+# 검색
 class ProductSearchView(generics.ListAPIView):
     serializer_class = ProductSearchSerializer
     permission_classes = [AllowAny]
-    pagination_class = PageNumberPagination
 
-    def get_queryset(self):
-        queryset = Product.objects.all()
-        
-        keyword = self.request.query_params.get('keyword', '')
-        if keyword:
-            queryset = queryset.filter(
-                Q(name__icontains=keyword) | 
-                Q(company__icontains=keyword) |
-                Q(ingredients__ingredient__name__icontains=keyword)
-            ).distinct()
-        
-        return queryset.order_by('-id')
+    class ListPagination(PageNumberPagination):
+        page_size = 10
+        page_query_param = "page"
+        page_size_query_param = None
+        max_page_size = 10
+
+    pagination_class = ListPagination
 
     @extend_schema(
         summary="제품 검색",
-        tags=["제품"],
         parameters=[
             OpenApiParameter(
-                name='keyword',
+                name="word",
                 type=OpenApiTypes.STR,
                 location=OpenApiParameter.QUERY,
-                description='검색 키워드'
-            )
-        ]
-    )
-    def get(self, request, *args, **kwargs):
-        return self.list(request, *args, **kwargs)
-
-# 메인 페이지 (GET /products/main/)
-class MainView(generics.ListAPIView):
-    serializer_class = MainSerializer
-    permission_classes = [AllowAny]
-
-    def get_queryset(self):
-        today = timezone.now().date()
-        start_date = today - timedelta(days=7)
-
-        # 최근 7일간 인기 제품
-        popular_products = Product.objects.annotate(
-            recent_views=Coalesce(
-                Sum('daily_views__views', filter=Q(daily_views__date__gte=start_date)),
-                0
-            )
-        ).order_by('-recent_views')[:8]
-
-        return popular_products
-
-    @extend_schema(
-        summary="메인 페이지 조회",
+                description="검색어(제품명, 회사명, 기능성 원료명)",
+                required=False,
+            ),
+            OpenApiParameter(
+                name="sort",
+                type=OpenApiTypes.STR,
+                location=OpenApiParameter.QUERY,
+                description="정렬: monthly_rank | price_asc | price_desc | review_desc (기본: monthly_rank)",
+                required=False,
+                enum=["monthly_rank", "price_asc", "price_desc", "review_desc"],
+            ),
+        ],
         tags=["제품"]
     )
     def get(self, request, *args, **kwargs):
         return self.list(request, *args, **kwargs)
 
-# 제품 이미지 업로드 (POST /products/<id>/images/)
+    def get_queryset(self):
+        sort = self.request.query_params.get("sort", "monthly_rank")
+        query = self.request.query_params.get("word")
+
+        today = timezone.now().date()
+        start_date = today - timedelta(days=30)
+
+        qs = Product.objects.all()
+
+        if query:
+            ingredient_product_ids = ProductIngredient.objects.filter(
+                Q(ingredient__name__icontains=query) | Q(ingredient__englishIngredient__icontains=query)
+            ).values("product_id")
+
+            qs = qs.filter(
+                Q(name__icontains=query) |
+                Q(company__icontains=query) |
+                Q(id__in=Subquery(ingredient_product_ids))
+            ).distinct()
+
+        if sort == "price_asc":
+            return qs.order_by("price", "id")
+
+        if sort == "price_desc":
+            return qs.order_by("-price", "id")
+
+        if sort == "review_desc":
+            return (
+                qs.annotate(review_count=Count("reviews", distinct=True))
+                .order_by("-review_count", "id")
+            )
+
+        # default: monthly_rank (최근 30일 조회수 합계 기준)
+        # 조회수 없는 상품도 포함(0으로 취급)하여 정렬
+        return (
+            qs.annotate(
+                totalViews=Coalesce(
+                    Sum(
+                        "daily_views__views",
+                        filter=Q(daily_views__date__gte=start_date) & Q(daily_views__date__lt=today),
+                    ),
+                    0,
+                )
+            )
+            .order_by("-totalViews", "id")
+        )
+
+# 메인
+class MainView(APIView):
+    permission_classes = [AllowAny]
+
+    @extend_schema(
+        summary="메인 화면",
+        tags=["제품"]
+    )
+    def get(self, request, *args, **kwargs):
+        # 소분류별 지난 30일 조회수 합계 순 상위 소분류 목록
+        today = timezone.now().date()
+        start_date = today - timedelta(days=30)
+
+        from products.models import SmallCategory
+
+        small_with_views = (
+            SmallCategory.objects
+            .exclude(category="전체")
+            .annotate(
+                totalViews=Coalesce(
+                    Sum(
+                        "category_products__product__daily_views__views",
+                        filter=Q(category_products__product__daily_views__date__gte=start_date)
+                    ),
+                    0,
+                )
+            )
+            .order_by("-totalViews", "id")
+        )
+
+        # 오늘 조회수 상위 10개 제품
+        top_today_products = (
+            Product.objects.annotate(
+                todayViews=Coalesce(
+                    Sum("daily_views__views", filter=Q(daily_views__date=today)),
+                    0,
+                )
+            )
+            .order_by("-todayViews", "id")[:10]
+        )
+
+        categories_payload = [
+            {"smallCategory": sc.category, "bigCategory": sc.bigCategory.category}
+            for sc in small_with_views
+        ]
+
+        serializer = MainSerializer(top_today_products, many=True)
+
+        return Response({
+            "topSmallCategories": categories_payload,
+            "topProductsToday": serializer.data,
+        })
+
+# 제품 이미지 등록
 class UploadProductImageView(APIView):
     permission_classes = [AllowAny]
     parser_classes = [parsers.MultiPartParser, parsers.FormParser]
 
     @extend_schema(
-        summary="제품 이미지 업로드",
+        summary="제품 이미지 등록",
         tags=["제품"]
     )
     def post(self, request, id):
@@ -200,6 +462,7 @@ class UploadProductImageView(APIView):
         except Product.DoesNotExist:
             raise NotFound("제품을 찾을 수 없습니다.")
         
+        # 여러 이미지 파일 받기
         images = request.FILES.getlist('images')
         
         if not images:
@@ -210,397 +473,3 @@ class UploadProductImageView(APIView):
         ProductImage.objects.bulk_create(uploaded_images)
         
         return Response({"success": True, "message": f"{len(uploaded_images)}개의 이미지가 등록되었습니다."}, status=201)
-
-# BigCategory 입력 화면 (템플릿 기반)
-def big_category_form(request):
-    """BigCategory 데이터 입력 화면"""
-    
-    if request.method == 'POST':
-        category = request.POST.get('category', '').strip()
-        
-        if category:
-            BigCategory.objects.create(category=category)
-            messages.success(request, f'"{category}" 카테고리가 성공적으로 추가되었습니다.')
-            return redirect('admin_big_category_form')
-        else:
-            messages.error(request, '카테고리 이름을 입력해주세요.')
-    
-    # 기존 카테고리 목록 가져오기
-    categories = BigCategory.objects.all().order_by('id')
-    
-    return render(request, 'products/big_category_form.html', {
-        'categories': categories
-    })
-
-# SmallCategory 입력 화면 (템플릿 기반)
-def small_category_form(request):
-    """SmallCategory 데이터 입력 화면"""
-    
-    if request.method == 'POST':
-        big_category_id = request.POST.get('bigCategory', '').strip()
-        category = request.POST.get('category', '').strip()
-        
-        if not big_category_id:
-            messages.error(request, '대분류를 선택해주세요.')
-        elif not category:
-            messages.error(request, '소분류 이름을 입력해주세요.')
-        else:
-            try:
-                big_category = BigCategory.objects.get(id=big_category_id)
-                SmallCategory.objects.create(
-                    bigCategory=big_category,
-                    category=category
-                )
-                messages.success(request, f'"{big_category.category} - {category}" 소분류가 성공적으로 추가되었습니다.')
-                return redirect('admin_small_category_form')
-            except BigCategory.DoesNotExist:
-                messages.error(request, '선택한 대분류를 찾을 수 없습니다.')
-    
-    # 기존 데이터 가져오기
-    big_categories = BigCategory.objects.all().order_by('id')
-    small_categories = SmallCategory.objects.select_related('bigCategory').all().order_by('bigCategory__id', 'id')
-    
-    return render(request, 'products/small_category_form.html', {
-        'big_categories': big_categories,
-        'small_categories': small_categories
-    })
-
-# Product 입력 화면 (템플릿 기반)
-def product_form(request):
-    """Product 데이터 입력 화면"""
-    
-    if request.method == 'POST':
-        # 기본 정보
-        name = request.POST.get('name', '').strip()
-        company = request.POST.get('company', '').strip()
-        price = request.POST.get('price', '').strip()
-        unit = request.POST.get('unit', '').strip()
-        piece = request.POST.get('piece', '').strip()
-        product_type = request.POST.get('productType', '').strip()
-        
-        # 필수 필드 검증
-        if not all([name, company, price, unit, piece, product_type]):
-            messages.error(request, '모든 필수 항목을 입력해주세요.')
-        else:
-            try:
-                # Product 생성
-                product = Product.objects.create(
-                    name=name,
-                    company=company,
-                    price=int(price),
-                    unit=unit,
-                    piece=piece,
-                    productType=product_type
-                )
-                
-                # 이미지 URL들 추가
-                image_urls = request.POST.getlist('image_urls[]')
-                for url in image_urls:
-                    url = url.strip()
-                    if url:
-                        ProductImage.objects.create(product=product, url=url)
-                
-                # 성분들 추가
-                ingredient_ids = request.POST.getlist('ingredient_ids[]')
-                amounts = request.POST.getlist('amounts[]')
-                for ingredient_id, amount in zip(ingredient_ids, amounts):
-                    ingredient_id = ingredient_id.strip()
-                    amount = amount.strip()
-                    if ingredient_id and amount:
-                        try:
-                            ingredient = Ingredient.objects.get(id=ingredient_id)
-                            ProductIngredient.objects.create(
-                                product=product,
-                                ingredient=ingredient,
-                                amount=amount
-                            )
-                        except Ingredient.DoesNotExist:
-                            pass
-                
-                # 카테고리들 추가
-                category_ids = request.POST.getlist('category_ids[]')
-                for category_id in category_ids:
-                    category_id = category_id.strip()
-                    if category_id:
-                        try:
-                            category = SmallCategory.objects.get(id=category_id)
-                            CategoryProduct.objects.create(
-                                product=product,
-                                category=category
-                            )
-                        except SmallCategory.DoesNotExist:
-                            pass
-                
-                messages.success(request, f'"{name}" 제품이 성공적으로 추가되었습니다.')
-                return redirect('admin_product_form')
-            except ValueError:
-                messages.error(request, '가격은 숫자로 입력해주세요.')
-            except Exception as e:
-                messages.error(request, f'오류가 발생했습니다: {str(e)}')
-    
-    # 기존 데이터 가져오기
-    ingredients = Ingredient.objects.all().order_by('name')
-    small_categories = SmallCategory.objects.select_related('bigCategory').all().order_by('bigCategory__id', 'id')
-    # 제품 정보를 더 자세히 가져오기 (이미지, 성분, 카테고리 개수 포함)
-    products = Product.objects.prefetch_related(
-        'images',
-        'ingredients',
-        'category_products__category__bigCategory'
-    ).annotate(
-        image_count=Count('images'),
-        ingredient_count=Count('ingredients')
-    ).order_by('-id')[:20]  # 최근 20개만 표시
-    
-    return render(request, 'products/product_form.html', {
-        'ingredients': ingredients,
-        'small_categories': small_categories,
-        'products': products
-    })
-
-# Ingredient 입력 화면 (템플릿 기반)
-def ingredient_form(request):
-    """Ingredient 데이터 입력 화면"""
-    
-    if request.method == 'POST':
-        name = request.POST.get('name', '').strip()
-        english_ingredient = request.POST.get('englishIngredient', '').strip()
-        min_recommended = request.POST.get('minRecommended', '').strip()
-        max_recommended = request.POST.get('maxRecommended', '').strip()
-        effect = request.POST.get('effect', '').strip()
-        side_effect = request.POST.get('sideEffect', '').strip()
-        
-        # 필수 필드 검증
-        if not all([name, english_ingredient, min_recommended, max_recommended, effect]):
-            messages.error(request, '필수 항목을 모두 입력해주세요.')
-        else:
-            Ingredient.objects.create(
-                name=name,
-                englishIngredient=english_ingredient,
-                minRecommended=min_recommended,
-                maxRecommended=max_recommended,
-                effect=effect,
-                sideEffect=side_effect if side_effect else None
-            )
-            messages.success(request, f'"{name}" 성분이 성공적으로 추가되었습니다.')
-            return redirect('admin_ingredient_form')
-    
-    # 기존 성분 목록 가져오기
-    ingredients = Ingredient.objects.all().order_by('name')
-    
-    return render(request, 'products/ingredient_form.html', {
-        'ingredients': ingredients
-    })
-
-# BigCategory 삭제
-def big_category_delete(request, category_id):
-    """BigCategory 삭제"""
-    try:
-        category = BigCategory.objects.get(id=category_id)
-    except BigCategory.DoesNotExist:
-        messages.error(request, '대분류를 찾을 수 없습니다.')
-        return redirect('admin_big_category_form')
-    
-    if request.method == 'POST':
-        category_name = category.category
-        # 연결된 SmallCategory의 CategoryProduct 먼저 삭제
-        small_categories = category.smallCategories.all()
-        for small_cat in small_categories:
-            CategoryProduct.objects.filter(category=small_cat).delete()
-        # BigCategory 삭제 (CASCADE로 SmallCategory도 자동 삭제됨)
-        category.delete()
-        messages.success(request, f'"{category_name}" 대분류가 성공적으로 삭제되었습니다.')
-        return redirect('admin_big_category_form')
-    
-    return render(request, 'products/big_category_delete_confirm.html', {
-        'category': category
-    })
-
-# SmallCategory 삭제
-def small_category_delete(request, category_id):
-    """SmallCategory 삭제"""
-    try:
-        category = SmallCategory.objects.get(id=category_id)
-    except SmallCategory.DoesNotExist:
-        messages.error(request, '소분류를 찾을 수 없습니다.')
-        return redirect('admin_small_category_form')
-    
-    if request.method == 'POST':
-        category_name = category.category
-        # PROTECT된 CategoryProduct 먼저 삭제
-        CategoryProduct.objects.filter(category=category).delete()
-        # SmallCategory 삭제
-        category.delete()
-        messages.success(request, f'"{category_name}" 소분류가 성공적으로 삭제되었습니다.')
-        return redirect('admin_small_category_form')
-    
-    return render(request, 'products/small_category_delete_confirm.html', {
-        'category': category
-    })
-
-# Ingredient 삭제
-def ingredient_delete(request, ingredient_id):
-    """Ingredient 삭제"""
-    try:
-        ingredient = Ingredient.objects.get(id=ingredient_id)
-    except Ingredient.DoesNotExist:
-        messages.error(request, '성분을 찾을 수 없습니다.')
-        return redirect('admin_ingredient_form')
-    
-    if request.method == 'POST':
-        ingredient_name = ingredient.name
-        # PROTECT된 ProductIngredient 먼저 삭제
-        ProductIngredient.objects.filter(ingredient=ingredient).delete()
-        # Ingredient 삭제
-        ingredient.delete()
-        messages.success(request, f'"{ingredient_name}" 성분이 성공적으로 삭제되었습니다.')
-        return redirect('admin_ingredient_form')
-    
-    return render(request, 'products/ingredient_delete_confirm.html', {
-        'ingredient': ingredient
-    })
-
-# Product 수정 화면 (템플릿 기반)
-def product_edit(request, product_id):
-    """Product 데이터 수정 화면"""
-    
-    try:
-        product = Product.objects.prefetch_related(
-            'images',
-            'ingredients__ingredient',
-            'category_products__category__bigCategory'
-        ).get(id=product_id)
-    except Product.DoesNotExist:
-        messages.error(request, '제품을 찾을 수 없습니다.')
-        return redirect('product_form')
-    
-    if request.method == 'POST':
-        # 기본 정보 수정
-        product.name = request.POST.get('name', '').strip()
-        product.company = request.POST.get('company', '').strip()
-        product.unit = request.POST.get('unit', '').strip()
-        product.piece = request.POST.get('piece', '').strip()
-        product.productType = request.POST.get('productType', '').strip()
-        
-        try:
-            product.price = int(request.POST.get('price', 0))
-        except ValueError:
-            messages.error(request, '가격은 숫자로 입력해주세요.')
-            return redirect('product_edit', product_id=product_id)
-        
-        # 필수 필드 검증
-        if not all([product.name, product.company, product.unit, product.piece, product.productType]):
-            messages.error(request, '모든 필수 항목을 입력해주세요.')
-        else:
-            product.save()
-            
-            # 기존 이미지 삭제 처리
-            delete_image_ids = request.POST.getlist('delete_image_ids[]')
-            if delete_image_ids:
-                ProductImage.objects.filter(id__in=delete_image_ids, product=product).delete()
-            
-            # 새 이미지 URL 추가
-            new_image_urls = request.POST.getlist('new_image_urls[]')
-            for url in new_image_urls:
-                url = url.strip()
-                if url:
-                    ProductImage.objects.create(product=product, url=url)
-            
-            # 기존 성분 삭제 처리
-            delete_ingredient_ids = request.POST.getlist('delete_ingredient_ids[]')
-            if delete_ingredient_ids:
-                ProductIngredient.objects.filter(id__in=delete_ingredient_ids, product=product).delete()
-            
-            # 기존 성분 수정
-            existing_product_ingredient_ids = request.POST.getlist('existing_product_ingredient_ids[]')
-            existing_ingredient_ids = request.POST.getlist('existing_ingredient_ids[]')
-            existing_amounts = request.POST.getlist('existing_amounts[]')
-            for prod_ing_id, ing_id, amount in zip(existing_product_ingredient_ids, existing_ingredient_ids, existing_amounts):
-                prod_ing_id = prod_ing_id.strip()
-                ing_id = ing_id.strip()
-                amount = amount.strip()
-                if prod_ing_id and ing_id and amount:
-                    try:
-                        product_ing = ProductIngredient.objects.get(id=prod_ing_id, product=product)
-                        ingredient = Ingredient.objects.get(id=ing_id)
-                        product_ing.ingredient = ingredient
-                        product_ing.amount = amount
-                        product_ing.save()
-                    except (ProductIngredient.DoesNotExist, Ingredient.DoesNotExist):
-                        pass
-            
-            # 새 성분 추가
-            new_ingredient_ids = request.POST.getlist('new_ingredient_ids[]')
-            new_amounts = request.POST.getlist('new_amounts[]')
-            for ingredient_id, amount in zip(new_ingredient_ids, new_amounts):
-                ingredient_id = ingredient_id.strip()
-                amount = amount.strip()
-                if ingredient_id and amount:
-                    try:
-                        ingredient = Ingredient.objects.get(id=ingredient_id)
-                        ProductIngredient.objects.create(
-                            product=product,
-                            ingredient=ingredient,
-                            amount=amount
-                        )
-                    except Ingredient.DoesNotExist:
-                        pass
-            
-            # 기존 카테고리 삭제 처리
-            delete_category_ids = request.POST.getlist('delete_category_ids[]')
-            if delete_category_ids:
-                CategoryProduct.objects.filter(id__in=delete_category_ids, product=product).delete()
-            
-            # 새 카테고리 추가
-            new_category_ids = request.POST.getlist('new_category_ids[]')
-            for category_id in new_category_ids:
-                category_id = category_id.strip()
-                if category_id:
-                    try:
-                        category = SmallCategory.objects.get(id=category_id)
-                        # 중복 체크
-                        if not CategoryProduct.objects.filter(product=product, category=category).exists():
-                            CategoryProduct.objects.create(
-                                product=product,
-                                category=category
-                            )
-                    except SmallCategory.DoesNotExist:
-                        pass
-            
-            messages.success(request, f'"{product.name}" 제품이 성공적으로 수정되었습니다.')
-            return redirect('admin_product_form')
-    
-    # 기존 데이터 가져오기
-    ingredients = Ingredient.objects.all().order_by('name')
-    small_categories = SmallCategory.objects.select_related('bigCategory').all().order_by('bigCategory__id', 'id')
-    
-    return render(request, 'products/product_edit.html', {
-        'product': product,
-        'ingredients': ingredients,
-        'small_categories': small_categories
-    })
-
-# Product 삭제
-def product_delete(request, product_id):
-    """Product 삭제"""
-    
-    try:
-        product = Product.objects.get(id=product_id)
-    except Product.DoesNotExist:
-        messages.error(request, '제품을 찾을 수 없습니다.')
-        return redirect('product_form')
-    
-    if request.method == 'POST':
-        product_name = product.name
-        
-        # PROTECT로 설정된 CategoryProduct를 먼저 삭제
-        CategoryProduct.objects.filter(product=product).delete()
-        
-        # 이제 제품 삭제 가능
-        product.delete()
-        messages.success(request, f'"{product_name}" 제품이 성공적으로 삭제되었습니다.')
-        return redirect('admin_product_form')
-    
-    # GET 요청 시 확인 페이지 표시
-    return render(request, 'products/product_delete_confirm.html', {
-        'product': product
-    })
